scripts/code_mixed_mt/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==0.21.0
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
transformers==4.35.2
tokenizers==0.15.0
accelerate==0.25.0
huggingface-hub==0.19.4